+++ bundles/customer-password-updated-at-connector/src/FondOfKudu/Client/CustomerPasswordUpdatedAtConnector/CustomerPasswordUpdatedAtConnectorClient.php
