+++ bundles/customer-password-updated-at-connector/src/FondOfKudu/Client/CustomerPasswordUpdatedAtConnector/CustomerPasswordUpdatedAtConnectorClientInterface.php
