+++ bundles/customer-password-updated-at-connector/src/FondOfKudu/Client/CustomerPasswordUpdatedAtConnector/Zed/CustomerPasswordUpdatedAtConnectorStub.php
