+++ bundles/customer-password-updated-at-connector/src/FondOfKudu/Shared/CustomerPasswordUpdatedAtConnector/CustomerPasswordUpdatedAtConnectorConstants.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace FondOfKudu\Shared\CustomerPasswordChangedAtConnector;
+namespace FondOfKudu\Shared\CustomerPasswordUpdatedAtConnector;
 
 interface CustomerPasswordUpdatedAtConnectorConstants
 {
