+++ bundles/customer-password-updated-at-connector/src/FondOfKudu/Shared/CustomerPasswordUpdatedAtConnector/Transfer/customer_password_updated_at_connector.transfer.xml
