+++ bundles/customer-password-updated-at-connector/src/FondOfKudu/Zed/CustomerPasswordUpdatedAtConnector/Business/CustomerExpander/CustomerExpanderInterface.php
