+++ bundles/customer-password-updated-at-connector/src/FondOfKudu/Zed/CustomerPasswordUpdatedAtConnector/CustomerPasswordUpdatedAtConnectorConfig.php
