+++ bundles/customer-password-updated-at-connector/src/FondOfKudu/Zed/CustomerPasswordUpdatedAtConnector/Dependency/QueryContainer/CustomerPasswordUpdatedAtConnectorToCustomerQueryContainerInterface.php
