+++ bundles/customer-password-updated-at-connector/src/FondOfKudu/Zed/CustomerPasswordUpdatedAtConnector/Persistence/Propel/Schema/spy_customer.schema.xml
